Profile is now shell-agnostic.
I primarily use zsh these days, but still use bash on some machines.

On machines where I use bash, I link this profile as ~/.bash_profile.

On machines where I use zsh, I link this profile as ~/.zprofile.

The file sets a few useful system variables, runs a bootstrap shell
script in my home directory if it exists (I use this to build my
environment, regardless of shell), and -- if this is a bash shell --
sources my bashrc. (If this is a zsh shell, the zshrc will be loaded
automatically and does not need to be sourced in ~/.zprofile.)