Remove abook alias
Mutt seems to not recognize this alias. As a result, I would have to add
the --datafile option to the abook calls in my muttrc, which makes
having an alias less useful. It seems simpler to me to just link the
addressbook file from my documents into ~/.abook/ and call abook without
any options.