Add abook alias
Apparently, whenever abook modifies its database, it overwrites the
file. If the database is linked into ~/.abook, whenever it is modified
abook will replace the database link with the actual a file. So that
doesn't really work.

Now I'm back to using an abook alias to specify the location of the
database. Mutt still doesn't respect my aliases, so I have to specify
the database location in three different locations in muttrc, as well.
So much for DRY.