Improve mutt signature matching in vim
A signature separator (`--`) should have a blank space after it, but
some mail clients (such as gmail) appear to strip this. By matching zero
or more white space characters between the separator and the end of line
I can account for this.

The grouping of the `> ` characters is not currently being used, but it
could be followed by a `\+` to strip signatures on any level of the
quote chain.